Create README.md
This  code  aims to give an  answer  to  the  following   Logic:

Suppose you're given a set which originally contains numbers from 1 to n. Unfortunately, due to a data error, one of the numbers in the set got duplicated to another number in the set, which results in a repetition of one number and a loss of another number. Given an array nums representing the data status of this set after the error, find and return the number that occurs twice and the number that is missing in the form of an array.